Add names command file and rework generateNamesLoad; getCurrent still does not work
