+++ packages/db/src/loaders/commands/index.ts
@@ -1,1 +1,2 @@
 export { generateCompileLoad } from "./compile";
+export { generateNamesLoad } from "./names";
